+++ R/get_data.R
@@ -813,7 +813,21 @@ get_data.plm <- function(x, verbose = TRUE, ...) {
   mf <- stats::model.frame(x)
   model_terms <- find_variables(x, effects = "all", component = "all", flatten = TRUE)
   cn <- colnames(mf)
-  mf <- as.data.frame(lapply(mf, as.vector))
+  mf <- as.data.frame(lapply(mf, function(i) {
+    if (is.factor(i)) {
+      as.factor(i)
+    } else if (is.character(i)) {
+      as.character(i)
+    } else if (is.integer(i)) {
+      as.integer(i)
+    } else if (is.numeric(i)) {
+      as.numeric(i)
+    } else if (is.logical(i)) {
+      as.logical(i)
+    } else {
+      as.vector(i)
+    }
+  }))
   colnames(mf) <- clean_names(cn)
 
   # find index variables
@@ -1110,7 +1124,7 @@ get_data.glht <- function(x, ...) {
 get_data.averaging <- function(x, ...) {
   ml <- attributes(x)$modelList
   if (is.null(ml)) {
-    warning("Can't retrieve data. Please use `fit = TRUE` in `model.avg()`.", call. = FALSE)
+    format_warning("Can't retrieve data. Please use `fit = TRUE` in `model.avg()`.")
     return(NULL)
   }